Add game(), change playRound() to return numbers
Instead of returning strings, playRound() will now return a value
corresponding to the result (-1 for tie, 0 for loss, 1 for win). Makes
it easier to parse the results in game(), plus we can use them for
calculations.

Also change error reporting to console.error() instead of alert().
Cases are most likely useless if the functions' logic hold up, but error
messages are there in the unlikely event that something goes wrong.